[#1599] Factor Secret model methods into feature modules
