--- apps/api/v2/models/definitions/secret_definition.rb
@@ -1,51 +1,0 @@
-# apps/api/v2/models/definitions/secret_definition.rb
-
-module V2
-  class Secret < Familia::Horreum
-    feature :safe_dump
-    feature :expiration
-
-    default_expiration 7.days # default only, can be overridden at create time
-    prefix :secret
-
-    identifier_field :key
-
-    field :custid
-    field :state
-    field :value
-    field :key
-    field :metadata_key
-    field :value_encryption
-    field :lifespan
-    field :share_domain
-    field :verification
-    field :updated
-    field :created
-    field :truncated # boolean
-
-    counter :view_count, default_expiration: 14.days # out lives the secret itself
-
-    # NOTE: this field is a nullop. It's only populated if a value was entered
-    # into a hidden field which is something a regular person would not do.
-    field :token
-
-    @safe_dump_fields = [
-      { identifier: ->(obj) { obj.identifier } },
-      :key,
-      :state,
-      { secret_ttl: ->(m) { m.lifespan } },
-      :lifespan,
-      { shortkey: ->(m) { m.key.slice(0, 8) } },
-      { has_passphrase: ->(m) { m.has_passphrase? } },
-      { verification: ->(m) { m.verification? } },
-      { is_truncated: ->(m) { m.truncated? } },
-      :created,
-      :updated,
-    ]
-
-    def init
-      self.state ||= 'new'
-      self.key   ||= self.class.generate_id # rubocop:disable Naming/MemoizedInstanceVariableName
-    end
-  end
-end
